requirements-api.txt
requests
beautifulsoup4
python-dotenv
fastapi
uvicorn
ratelimit
selenium
webdriver-manager
